requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic[email]>=2.10.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0

